alerts checked
